personal_feed_project/requirements.txt
Django==4.2.7
django-filter==23.3
djangorestframework==3.14.0
drf-spectacular==0.26.5
Faker==19.13.0
flake8==6.1.0
inflection==0.5.1
isort==5.12.0
jsonschema==4.19.2
Pillow==10.1.0
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1